refactor(): Split into files
