+++ internal/processor/config_test.go
@@ -1,0 +1,24 @@
+package processor_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/oleg-balunenko/spamassassin-parser/internal/processor"
+)
+
+func TestNewConfig(t *testing.T) {
+	expConfgig := &processor.Config{
+		Buffer: 0,
+		Receive: struct {
+			Response bool
+			Errors   bool
+		}{
+			Response: true,
+			Errors:   false,
+		},
+	}
+	got := processor.NewConfig()
+	require.Equal(t, expConfgig, got)
+}
